Added F90-C compatibility flag for intel compiler
This flag ensures that, in F90, .TRUE. has an integer value of 1,
and .FALSE. an integer value of 0. This ensures that, if x is a
boolean returned from a F90 subroutine, the comparison x==true
in C will pass if x was set to .TRUE. in F90. Without this flag,
potentially x could contain the integer value of 255 (i.e., -1),
and make that comparison fail.

Note: I'm not sure why this flag is needed now, while it wasn't before.