Implements murphy and Koop formulation for computing saturation vapor pressure (#430)
Adds murphy+Koop saturation vapor pressure function and makes it the default choice for P3. This is more accurate at very cold and very warm temperatures. Supposedly more computationally expensive, but difference seems to be in the noise (as documented in this PR). Has mild effects on model climate. 

We still need to make Murphy+Koop the default saturation scheme for SHOC and need to improve its NaN handling. These will occur in a future PR.