Modified approach for op< overload for weak_ptr
Instead of overloading op< for weak_ptr<T>, we now have a compare struct,
which the user can pass as template arg for certain std containers,
such as std::set.

The issue with op< was the inability of the compiler to look it up.
Explicitly passing a Compare functor type to the template arguments
list of, say, std::set is easy, so we do that instead now.