Micro mod to kokkos and yaml-cpp build scripts.
When cmake is triggered (e.g., touching a CMakeLists.txt file), cmake would
find the [Kokkos|YAMLCPP]_SOURCE_DIR in the cache, and until now that triggered
the message warning that user-provided versions might not work.

Now, we check the provided path, and if it matches the submodule path,
then we do not scare the user like that.