Update kokkos submodule to fix libdl issue
This update switches the remote repo for kokkos from the kokkos/kokkos.git one
to E3SM-Project/kokkos. The latter is E3SM fork, where we are able to put in
hot fixes relatively quickly. This update, in particular, fixes an issue that
kokkos has when using '-static' linking on CRAY systems.