Add some overloads for volatile Packs
I uncovered the need for these overloads when using Packs inside Kokkos scan
operations. In particular, in Kokkos atomic fetch_add impl, there's some use
of volatile, which when applied to a pack makes binding to const ref illegal.

The solution is to add overloads that take a const volatile Pack ref.
I only added overloads for the needed methods. It is possible that other uses
of Pack might unveil the need for additional ones, but there's no need to
add all of them now. If need be, we'll add them later.