Add MemoryTraits to template signature of view aliases
For all the Kokkos:View template aliases declared in KokkosTypes,
allow to specify the memory traits. If no memory traits are specified,
default to Kokkos::MemoryManaged (which is what Kokkos does internally).

The reason is that if you don't specify MemoryTraits in a view template
signature, the resulting type is different (and therefore incompatible)
with a type that has it declared. E.g., with

 using V1 = Kokkos::View<int*,Kokkos::MemoryManaged>;
 using V2 = Kokkos::View<int*>;

an object of type V2 cannot be bound to a reference V2&. This makes
writing functions that take in, say, view_1d<Real>, more complicated.
E.g., the two functions

  void f1 (const view_1d<Real>& v);
  void f2 (const ekat::Unmanaged<view_1d<Real>>& v);

will both fail to compile if passed either a V2 or V1 respectively.
Adding MemoryTraits to the templated signature of view_1d makes
things easier:

  template<typename MT>
  void f1 (const view_1d<Real,MT>& v);

  view_2d<Real> a("",10,10
  view_1d<Real> b("",10);
  auto c = subivew(a,0); // this view has MemoryTraits=Kokkos::MemoryUnmanaged

  f1(b); // works
  f1(c); // works