Better handle the OpenMP dependency in ekat
Kokkos still relies on setting the flags. But that can be problematic.
E.g., Kokkos adds its link options to the target link properties only
if the link language is CXX (this helps avoiding cuda link options to
be passed on to a non-cuda linker). However, when the main entry point
is in F90, and the link language is Fortran, this causes the lack of
openmp flags in the link line, and consequently undef refs to omp stuff.

Instead, if kokkos enables openmp, we explicitly link ekat against the
openmp cmake target (for cxx and fortran). This will allow all compiler
flags and linker flags to be correctly propagated, without the need
to assume anything about the link/compiler language of the downstream
applications.