Removed definition of 'restrict' macro
Since it was not declared for Intel compiler, this macro was broken,
and using it would have created compilation errors on Intel.
Since it was unused, one could have kept it but, a) keeping it would
have given the false sense of security that it worked, and b) some
downstream package might want to declare a similar macro (the name
can easily clash).