Add test to ensure input redirection is not parsed as a filter
If the redirection of the input does not work as a redirection,
no tests are run, which mean the return code of the exec would be 0.
So tell ctest that the test *should fail*, so that if no tests are
run, ctest craps out (no error code returned). To ensure the exec
returns a non-zero return code, I added a REQUIRE(false) in the test.