Changed how EKAT handles MPI in order to build
Instead of using MPI compiler wrappers, we use native
C++/Fortran compilers, and link against MPI libs
(which is what the wrapper do for you in the background).

This allows to simplify and/or remove some cmake logic
in EKAT. Furthermore, it will easily allow, in a following
PR, to turn MPI into an optional dependency.