Changed how '--use-colour no' is added to unit tests
Since EXCLUDE_MAIN_CPP is not an option for the test phase,
the function EkatCreateUnitTestFromExec cannot use it to determine
whether the option is to be added. Instead, we require it to
be passed as part of EXE_ARGS. For the combo function
EkatCreateUnitTest, we have access to EXCLUDE_MAIN_CPP from the
EkatCreateUnitTestExec phase, and we can manually add it to EXE_ARGS
before calling EkatCreateUnitTestFromExec.