Adds EKAT_ENABLE_FPE option for enabling floating point exceptions.
This is an option we can use to decide whether we want EKAT to interact with
the floating point exception system at all. It's ON by default, to keep
behavior unchanged. Might come in handy on architectures where it's less
obvious how to handle floating point exception trapping (such as Mac M1).