Make sure the name of the log file is available on all ranks
If a file is generated at all (i.e., LogFilePolicy!=LofNoFile),
we want all ranks to know the name of the file, since they might
need it, even if they don't log anything. E.g., they might
need to create a second log file based on the name of the first.