Fix bug in TeamUtils get/release_workspace_idx for GPU.
We need a memory_fence so that read/writes to work slots are sequenced properly
w.r.t. the lock on the workspace slots.

Add a test that fails in ~2 out of 5 trials with the buggy code. Passes 100
times in a row with the fix.

Use Kokkos::single's broadcast instead of a separate reduction.