Refactor EKAT into packages
The packages are Core, Utils, TestUtils, Kokkos, Pack, Algorithm,
Units, YamlParser, Logging, and MpiComm. All except the first are
OPTIONAL, but there are, of course, dependencies.

Some of the existing Ekat code has been removed, since, to the best
of our knowledge, was not used. Some has been moved to a different
file, or had the types renamed.