Reintroduce invalid fcn but not as part of ScalarTraits
On CUDA (probably due to ODR) I can't make the
Kokkos::Experimental::xyz_v template variables
to work. Reverting back to an invalid method
works.

This util, which must be a KOKKOS_INLINE_FUNCTION, cannot be
a member of ScalarTraits, which must be usable even if Kokkos is NOT compiled